Fix for warnings
[FIX] Fixed warnings for unused function parameters whenever applicable
[FIX] Fixed type truncation warnings converting from larger to smaller types and vice versa
[FIX] Fixed warnings for uninitialized variables, also for the BYTE_PADDING macro
[FIX] Added defaults for structs when applicable and missing

+++ DX12Renderer/Source/Graphics/Backend/RootSignature.cpp
@@ -53,7 +53,7 @@ void RootSignature::Create(const std::string& name, const std::vector<CD3DX12_RO
 	staticSamplers[1].ShaderVisibility = D3D12_SHADER_VISIBILITY_PIXEL;
 
 	CD3DX12_VERSIONED_ROOT_SIGNATURE_DESC versionedRootSignatureDesc = {};
-	versionedRootSignatureDesc.Init_1_1(rootParameters.size(), &rootParameters[0], _countof(staticSamplers), &staticSamplers[0], rootSignatureFlags);
+	versionedRootSignatureDesc.Init_1_1(static_cast<uint32_t>(rootParameters.size()), &rootParameters[0], _countof(staticSamplers), &staticSamplers[0], rootSignatureFlags);
 
 	RenderBackend::GetDevice()->CreateRootSignature(versionedRootSignatureDesc, m_d3d12RootSignature);
 	m_d3d12RootSignature->SetName(StringHelper::StringToWString(name + " root signature").c_str());
